Fix percpu maps in syscall probes
Syscall probes cannot run in critical sections, as they can call
into functions that can block or sleep.  This has meant that there
was a race condition where a thread could fetch a percpu value
from a map, start using it and then block.

Another thread could then be scheduled on that cpu, fetch the
same percpu value and overwrite it.  The original thread would
potentially have garbage data when it began executing the EBPF
program again.

Fix this by having an sx lock for every CPU.  We take the lock
for the current CPU before entering the program, and then every
percpu lookup takes place as if we are still running on that
CPU (even if we have since migrated CPUs).  If a second thread
on the original CPU attemps to enter the program, it will block
on the sx lock until the first thread completes.

I'm ensure of the performance implications of all this but it
at least closes the race.