Update progtable array to compile a vm
When a progtable array has an element assigned, compile it down
to a VM.  This is pretty ugly and needs a significant rework.
It's likely that the current distinction between "program" and
"vm" is not useful and getting in the way, as the problem stems
from the fact that the struct file contains a program but not a
vm, meaning that the struct file cannot be easily run.