Rework progarray again
Malloc cannot be allocated in the ebpf_map_update() path.  Give
up on this and make a progarray be a simple array of file
descriptors that validate that a descriptor points to an EBPF
program.  It is the responsibility of any consumer of a progarray
to ensure that the descriptor is still valid and still points to
a program when the array is used.

ebpf_prog instances now have a ebpf_vm instance embedded in then
that can be used for running the code in the program.