Add a blocking pdwait4 EBPF function
Add an EBPF function that performs a pdwait4.  To avoid blocking
in EBPF probe context, which for wait4 could cause a deadlock,
instead the EBPF caller must provide an FD pointing to another
EBPF program.  The blocking pdwait4 call will be done out of
probe context, and that program will be called with the result.