Add EBPF function to canonicalize a path
Add an EBPF function that takes two paths and overwrites the
first with the path pointed to by the second, relative to the
first if the second is relative.

This handles removing .. and . path components.  An error is
returned if the caller tries to .. outside of the bounds of
the first path.