Preserve trailing slash in canonical paths
In certain cases, like when the final component of a path is a
symlink and AT_SYMLINK_NOFOLLOW is specified, there is a
difference in behaviour depending on whether the path ends in
a slash or not.  Preserve a single trailing slash when
canonicalizing a path so that we emulate this behaviour properly.