Make resolve_one_symlink obey AT_SYMLINK_NOFOLLOW
Pass flags to resolve_one_symlink and have it obey the
AT_SYMLINK_NOFOLLOW flag.  If this flag is specified, the last
component of a path will not be resolved if it is a symlink.
Adding the additional parameter has caused us to go over the
argument limit so the arguments have been reworked.  The first
argument is now a struct; the first component of the struct
contains the buffer with the base path in it, and there
are two scratch buffers.  This allows us to no longer allocate
a scratch buffer on the stack.